+++ src/commands/open-wiki.ts
@@ -1,76 +1,63 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as vscode from "vscode";
-import environment from "../environment";
-import * as github from "../github";
+import { Environment } from "../extension";
+import github from "../github";
 
-export default async function openWiki() {
-  // ask for user's GitHub username
-  let username = environment.config.get<string>("github.username");
-  if (!username) {
-    username = await vscode.window.showInputBox({
-      placeHolder: "Enter a GitHub username...",
-    });
-  }
-  if (!username) {
-    return;
-  }
+export default async function openWiki(env: Environment) {
+  // get user's github session
+  const session = await vscode.authentication.getSession("github", ["repo"], {
+    createIfNone: true,
+  });
 
-  // get all user's repo and ask them to select
-  const repos = await github.getRepos(username).then((repos) =>
-    repos.filter((repo) => {
-      // filter out repos that don't have wiki or are archived
-      return repo.has_wiki && !repo.archived;
-    })
-  );
-  const repo = await vscode.window
-    .showQuickPick(
-      repos.map((repo) => {
-        return {
-          label: repo.name,
-          description: repo.description,
-          data: repo,
-        };
-      }),
-      {
-        placeHolder: "Select a wiki to open...",
-      }
+  // get user's repos
+  github.updateToken(session?.accessToken);
+  let repos = await github.getUserRepos();
+  repos = repos.filter((repo) => {
+    return repo.has_wiki && !repo.archived;
+  });
+
+  // prompt user to select a wiki
+  const repo = (
+    await vscode.window.showQuickPick(
+      repos.map(
+        (repo) => {
+          return {
+            label: repo.full_name,
+            detail: repo.description,
+            iconPath: new vscode.ThemeIcon("github"),
+            data: repo,
+          };
+        },
+        {
+          placeHolder: "Select a wiki to open...",
+        }
+      )
     )
-    .then((item) => item?.data);
-
+  )?.data;
   if (!repo) {
     return;
   }
 
-  const wikiBasePath = path.join(environment.tempDir, repo.full_name);
+  const wikiBasePath = path.join(env.tempDir, repo.full_name);
   const wikiSourcePath = path.join(wikiBasePath, "wiki");
   const wikiWorkspacePath = path.join(wikiBasePath, "wiki.code-workspace");
 
-  // checks if wiki already exists
+  // check if wiki workspace already exists
   if (fs.existsSync(wikiWorkspacePath)) {
-    const newWorkspace = await vscode.window
-      .showInformationMessage(
-        "A wiki workspace already exists. Do you want to open it or create a new one?",
-        "Open",
-        "New"
-      )
-      .then(async (value) => {
-        if (value === "New") {
-          return true;
-        }
-        return false;
-      });
-    if (newWorkspace) {
-      // delete existing wiki workspace
-      fs.rmdirSync(wikiBasePath, { recursive: true });
-    } else {
-      // open existing wiki workspace
-      await vscode.commands.executeCommand(
+    const newWorkspace = await vscode.window.showInformationMessage(
+      "Wiki workspace already exists. Open a new workspace?",
+      "Yes",
+      "No"
+    );
+    if (newWorkspace !== "Yes") {
+      vscode.commands.executeCommand(
         "vscode.openFolder",
         vscode.Uri.file(wikiWorkspacePath)
       );
       return;
     }
+    fs.rmSync(wikiBasePath, { recursive: true });
   }
 
   // create wiki workspace
@@ -98,7 +85,7 @@ export default async function openWiki() {
   );
 
   // open wiki workspace
-  await vscode.commands.executeCommand(
+  vscode.commands.executeCommand(
     "vscode.openFolder",
     vscode.Uri.file(wikiWorkspacePath)
   );