+++ src/commands/publish-wiki.ts
@@ -1,47 +1,56 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as vscode from "vscode";
-import environment from "../environment";
-import * as utils from "../utils";
+import { Environment } from "../extension";
+import {
+    executeShellCommands,
+    executeTerminalCommands,
+    matchRepoFullName,
+} from "../utils";
 
-export default async function publishWiki() {
-  const isWikiWorkspace = environment.config.get<boolean>("workspace.isWikiWorkspace", false);
-  const repoFullName = environment.config.get<string>("workspace.repoFullName");
-
+export default async function publishWiki(env: Environment) {
+  // check if the current workspace is a valid wiki workspace
+  const isWikiWorkspace = env.config.get<boolean>(
+    "workspace.isWikiWorkspace",
+    false
+  );
+  const repoFullName = env.config.get<string>("workspace.repoFullName");
   if (!isWikiWorkspace) {
     vscode.window.showErrorMessage("Please open a wiki first!");
     return;
   }
   if (!repoFullName) {
-    vscode.window.showErrorMessage("Invalid wiki workspace.");
+    vscode.window.showErrorMessage("Invalid wiki workspace!");
+    return;
+  }
+  const wikiPath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
+  if (!wikiPath) {
+    vscode.window.showErrorMessage("Unable to find the wiki path!");
     return;
   }
 
-  const wikiPath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-  if (!wikiPath) {
-    vscode.window.showErrorMessage("Unable to find the wiki path.");
-    return;
-  }
+  // check if the wiki is initialized
   if (!fs.existsSync(path.join(wikiPath, ".git"))) {
     vscode.window.showErrorMessage("Please initialize the wiki first!");
     return;
   }
 
   // gets the remote repo's full name
-  const remoteRepoFullName = await utils
-    .executeShellCommands(`cd ${wikiPath} && git remote -v`)
-    .then(
-      (result) => {
-        const remoteRepoFullName = utils.matchRepoFullName(result);
-        console.log(remoteRepoFullName);
-        return remoteRepoFullName;
-      },
-      (err) => {
-        console.error(err);
-      }
+  const remoteRepoFullName = await executeShellCommands(
+    `cd ${wikiPath}`,
+    `git remote -v`
+  ).then(
+    (result) => matchRepoFullName(result),
+    (_) => undefined
+  );
+  if (!remoteRepoFullName) {
+    vscode.window.showErrorMessage(
+      "Unable to find the remote repo's full name."
     );
+    return;
+  }
 
-  // checks whether the local git origin matches the wiki's remote
+  // checks if the local git origin matches the wiki's remote
   if (remoteRepoFullName !== repoFullName) {
     vscode.window.showErrorMessage(
       "The git origin does not match the wiki's remote."
@@ -49,7 +58,7 @@ export default async function publishWiki() {
     return;
   }
 
-  // ask user to select mode of push
+  // ask for the mode of commit
   const orphanCommit = (
     await vscode.window.showQuickPick([
       {
@@ -59,7 +68,8 @@ export default async function publishWiki() {
       },
       {
         label: "Push as a orphaned commit",
-        description: "Sqush all your changes including your history into one commit.",
+        description:
+          "Sqush all your changes including your history into one commit.",
         data: true,
       },
     ])
@@ -71,16 +81,16 @@ export default async function publishWiki() {
   });
   if (!commitMessage) {
     vscode.window.showErrorMessage(
-      "Operation cancelled. Please enter a commit message!"
+      "Operation cancelled! Please enter a commit message."
     );
     return;
   }
 
   // push the changes to the wiki
-  utils.executeCommands(`cd ${wikiPath}`);
+  executeTerminalCommands(`cd ${wikiPath}`);
   if (orphanCommit) {
     // push as a orphaned commit
-    utils.executeCommands(
+    executeTerminalCommands(
       `git checkout --orphan wiki`,
       `git add .`,
       `git commit -m "${commitMessage}"`,
@@ -91,12 +101,12 @@ export default async function publishWiki() {
     );
   } else {
     // push as a new commit
-    utils.executeCommands(
+    executeTerminalCommands(
       `git add .`,
       `git commit -m "${commitMessage}"`,
       `git push origin master`
     );
-
-    vscode.window.showInformationMessage("Wiki published!");
   }
+
+  vscode.window.showInformationMessage("Wiki published!");
 }