+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this extension will be documented in this file.
 
+## [Unreleased]
+
+- Add GitHub authentication
+- Change repo description to detail in quick pick
+
 ## [1.0.0] - 2024-03-04
 
 The changes for this version is documented [here](https://github.com/dentolos19/git-wiki-editor/releases/tag/v1.0.0).